requirements.txt
gspread==6.*
google-auth==2.*
requests==2.*
python-dotenv==1.*
